backend/tests/requirements.txt
# Testing Dependencies for KPI Operations Backend

# Core Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
pytest-mock==3.12.0
pytest-xdist==3.5.0  # Parallel test execution

# Database Testing
pytest-postgresql==5.0.0
sqlalchemy-utils==0.41.1

# API Testing
httpx==0.25.2
requests-mock==1.11.0

# Fixtures and Factories
factory-boy==3.3.0
faker==20.1.0

# Coverage
coverage[toml]==7.3.4

# Assertions
assertpy==1.1

# Performance Testing
pytest-benchmark==4.0.0

# Test Reporting
pytest-html==4.1.1
pytest-json-report==1.5.0
